docs and rename of folder
