+++ API-Samples/Java/app/build.gradle
@@ -24,8 +24,6 @@ dependencies {
     implementation 'com.azure:azure-identity:1.8.0'
     implementation 'com.azure:azure-messaging-eventhubs:5.15.2'
     implementation 'com.azure:azure-messaging-eventhubs-checkpointstore-blob:1.16.3'
-
-
 }
 
 testing {
