+++ API-Samples/Java/app/src/main/java/apisamples/Account/Account.java
@@ -1,4 +1,4 @@
-package videoindexersamples.Account;
+package apisamples.Account;
 
 public class Account {
     public AccountProperties properties;
