+++ API-Samples/Java/app/src/main/java/apisamples/Account/AccountProperties.java
@@ -1,4 +1,4 @@
-package videoindexersamples.Account;
+package apisamples.Account;
 
 /**
  * For Full Properties List
