+++ API-Samples/Java/app/src/main/java/apisamples/ProcessingState.java
@@ -1,4 +1,4 @@
-package videoindexersamples;
+package apisamples;
 
 public enum ProcessingState {
     Uploaded,
