+++ API-Samples/Java/app/src/main/java/apisamples/Video.java
@@ -1,4 +1,4 @@
-package videoindexersamples;
+package apisamples;
 
 public class Video {
     public String id;
