+++ API-Samples/Java/app/src/main/java/apisamples/authentication/AccessTokenRequest.java
@@ -1,4 +1,4 @@
-package videoindexersamples.authentication;
+package apisamples.authentication;
 
 
 /**
