+++ API-Samples/Java/app/src/main/java/apisamples/authentication/AccessTokenResponse.java
@@ -1,4 +1,4 @@
-package videoindexersamples.authentication;
+package apisamples.authentication;
 
 public class AccessTokenResponse {
     public String accessToken;
