+++ API-Samples/Java/app/src/main/java/apisamples/authentication/ArmAccessTokenScope.java
@@ -1,4 +1,4 @@
-package videoindexersamples.authentication;
+package apisamples.authentication;
 
 public enum ArmAccessTokenScope {
     Account,
